feat: fixed bugs in the backend

+++ second/src/controllers/comment.controller.js
@@ -9,7 +9,7 @@ import { getIO } from "../config/socket.js"
 const getVideoComments = asyncHandler(async (req, res) => {
     const { videoId } = req.params;
 
-    // ✅ Validate ObjectId (VERY IMPORTANT)
+    // Validate ObjectId
     if (!mongoose.Types.ObjectId.isValid(videoId)) {
         return res.status(400).json({
             success: false,
@@ -17,7 +17,7 @@ const getVideoComments = asyncHandler(async (req, res) => {
         });
     }
 
-    // ✅ Fetch comments
+    // Fetch comments
     const comments = await Comment.find({ video: videoId })
         .sort({ createdAt: -1 }) // newest first
         .populate("owner", "username avatar")
@@ -30,7 +30,6 @@ const getVideoComments = asyncHandler(async (req, res) => {
 });
 
 const addComment = asyncHandler(async (req, res) => {
-    // TODO: add a comment to a video
     const { videoId } = req.params
     const { content } = req.body
     const comment = await Comment.create(
@@ -58,25 +57,20 @@ const addComment = asyncHandler(async (req, res) => {
 })
 
 const updateComment = asyncHandler(async (req, res) => {
-    // TODO: update a comment
     const { commentId } = req.params
     const { content } = req.body
-    const updatedComment = await Comment.findByIdAndUpdate(
-        commentId,
 
-        {
-            $set: {
-                content: content
-            }
-
-        },
-        {
-            new: true
-        }
-    )
-    if (!updatedComment) {
+    const comment = await Comment.findById(commentId)
+    if (!comment) {
         throw new ApiError(404, "Comment not found")
     }
+
+    if (comment.owner.toString() !== req.user._id.toString()) {
+        throw new ApiError(403, "You are not authorized to update this comment")
+    }
+
+    comment.content = content
+    const updatedComment = await comment.save()
 
     return res
         .status(200)
@@ -84,16 +78,22 @@ const updateComment = asyncHandler(async (req, res) => {
 })
 
 const deleteComment = asyncHandler(async (req, res) => {
-    // TODO: delete a comment
     const { commentId } = req.params
-    const deletedComment = await Comment.findByIdAndDelete(commentId)
-    if (!deletedComment) {
+
+    const comment = await Comment.findById(commentId)
+    if (!comment) {
         throw new ApiError(404, "Comment not found")
     }
+
+    if (comment.owner.toString() !== req.user._id.toString()) {
+        throw new ApiError(403, "You are not authorized to delete this comment")
+    }
+
+    const deletedComment = await Comment.findByIdAndDelete(commentId)
+    
     return res
         .status(200)
         .json(new ApiResponse(200, deletedComment, "Comment deleted Successfully"))
-
 })
 
 export {
